+++ packages/gluestick/src/renderer/response/getHeaders.js
