+++ packages/gluestick/src/renderer/response/setHeaders.js
